Dump stack traces.

+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/image/AOTSamplingData.java
@@ -2,5 +2,5 @@ package com.oracle.svm.hosted.image;
 
 public interface AOTSamplingData {
 
-    void addEntry(int adr, int methodId);
+    void addEntry(long adr, int methodId);
 }
